Move quic-client to its own crate
