+++ client/src/nonblocking/quic_client.rs
@@ -1,7 +1,7 @@
 //! Simple nonblocking client that connects to a given UDP port with the QUIC protocol
 //! and provides an interface for sending transactions which is restricted by the
 //! server's flow control.
-pub use solana_tpu_client::nonblocking::quic_client::{
+pub use solana_quic_client::nonblocking::quic_client::{
     QuicClient, QuicClientCertificate, QuicError, QuicLazyInitializedEndpoint, QuicTpuConnection,
 };
 use {
