+++ quic-client/Cargo.toml
@@ -10,11 +10,27 @@ documentation = "https://docs.rs/solana-quic-client"
 edition = "2021"
 
 [dependencies]
+async-mutex = "1.4.0"
+async-trait = "0.1.57"
+futures = "0.3"
+itertools = "0.10.5"
+lazy_static = "1.4.0"
+log = "0.4.17"
+quinn = "0.8.4"
+quinn-proto = "0.8.4"
+quinn-udp = "0.1.3"
+rustls = { version = "0.20.6", features = ["dangerous_configuration"] }
+solana-measure = { path = "../measure", version = "=1.15.0" }
 solana-metrics = { path = "../metrics", version = "=1.15.0" }
+solana-net-utils = { path = "../net-utils", version = "=1.15.0" }
+solana-rpc-client-api = { path = "../rpc-client-api", version = "=1.15.0" }
 solana-sdk = { path = "../sdk", version = "=1.15.0" }
 solana-streamer = { path = "../streamer", version = "=1.15.0" }
 solana-tpu-client = { path = "../tpu-client", version = "=1.15.0" }
 thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
 
 [dev-dependencies]
+crossbeam-channel = "0.5"
 solana-logger = { path = "../logger", version = "=1.15.0" }
+solana-perf = { path = "../perf", version = "=1.15.0" }
