+++ quic-client/src/lib.rs
@@ -3,6 +3,9 @@
 pub mod nonblocking;
 pub mod quic_client;
 
+#[macro_use]
+extern crate solana_metrics;
+
 use {
     crate::{
         nonblocking::quic_client::{
