+++ tpu-client/src/lib.rs
@@ -2,7 +2,6 @@
 
 pub mod connection_cache_stats;
 pub mod nonblocking;
-pub mod quic_client;
 pub mod tpu_client;
 pub mod tpu_connection;
 pub mod tpu_connection_cache;
