+++ tpu-client/src/nonblocking/mod.rs
@@ -1,3 +1,2 @@
-pub mod quic_client;
 pub mod tpu_client;
 pub mod tpu_connection;
